Starting work on single venue page
Starting work on single venue page. Lots of work yet to be done, but a decent start.

I got all my data and I only need to create the HTML and CSS now.

One challenge will be a proper image carousel, as well as handling instances where there are no nested objects in the media object. Will get to that later though